Move table tests to JDBC module
