+++ stdlib/jdbc/src/test/resources/testng.xml
@@ -36,6 +36,9 @@ under the License.
             <class name="org.ballerinax.jdbc.SQLActionsTest"/>
             <class name="org.ballerinax.jdbc.SQLConnectionPoolTest"/>
             <class name="org.ballerinax.jdbc.SQLConnectorInitTest"/>
+
+<!--            <class name="org.ballerinax.jdbc.table.TableTest"/>-->
+            <class name="org.ballerinax.jdbc.table.TableIterationTest"/>
         </classes>
     </test>
 </suite>
